backend/requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-environ==0.11.2
django-extensions==3.2.3
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
openpyxl==3.1.2
Pillow==10.1.0
gunicorn==21.2.0
whitenoise==6.6.0
